manutencaoVendas

+++ nds-repository-client/src/main/java/br/com/abril/nds/repository/EstoqueProdutoCotaRepository.java
@@ -50,5 +50,5 @@ public interface EstoqueProdutoCotaRepository extends Repository<EstoqueProdutoC
 	public abstract ItemDTO<Long, BigInteger> loadIdAndQtdDevolvidaByIdProdutoEdicaoAndIdCota(
 			final Long idProdutoEdicao, final Long idCota);
 
-	BigInteger obterVendaBaseadoNoEstoque(Long idProdutoEdicao);
+	BigDecimal obterVendaBaseadoNoEstoque(Long idProdutoEdicao);
 }
